chore: updated the app dir folder structure to achieve the authotization
