+++ app/(auth)/login/_components/Login.tsx
@@ -1,0 +1,33 @@
+'use client';
+
+import Link from 'next/link';
+
+import LoginView from '@/app/_views/LoginView';
+import isAuth from '@/components/auth/is-auth';
+import { Icons } from '@/components/icons';
+import { buttonVariants } from '@/components/ui/button';
+import { cn } from '@/lib/utils';
+
+const Login = () => {
+  return (
+    <div>
+      <Link
+        href='/'
+        className={cn(
+          buttonVariants({ variant: 'ghost' }),
+          'absolute left-4 top-4 md:left-8 md:top-8'
+        )}
+      >
+        <>
+          <Icons.chevronLeft className='mr-2 h-4 w-4' />
+          Back
+        </>
+      </Link>
+      <div className='mx-auto flex w-full flex-col justify-center space-y-6 sm:w-[350px]'>
+        <LoginView />
+      </div>
+    </div>
+  );
+};
+
+export default isAuth(Login);
