+++ app/(auth)/login/page.tsx
@@ -1,10 +1,6 @@
 import { Metadata } from 'next';
-import Link from 'next/link';
 
-import LoginView from '@/app/_views/LoginView';
-import { Icons } from '@/components/icons';
-import { buttonVariants } from '@/components/ui/button';
-import { cn } from '@/lib/utils';
+import Login from './_components/Login';
 
 export const metadata: Metadata = {
   title: 'Login',
@@ -12,23 +8,5 @@ export const metadata: Metadata = {
 };
 
 export default function LoginPage() {
-  return (
-    <div className='container flex h-screen w-screen flex-col items-center justify-center'>
-      <Link
-        href='/'
-        className={cn(
-          buttonVariants({ variant: 'ghost' }),
-          'absolute left-4 top-4 md:left-8 md:top-8'
-        )}
-      >
-        <>
-          <Icons.chevronLeft className='mr-2 h-4 w-4' />
-          Back
-        </>
-      </Link>
-      <div className='mx-auto flex w-full flex-col justify-center space-y-6 sm:w-[350px]'>
-        <LoginView />
-      </div>
-    </div>
-  );
+  return <Login />;
 }
