+++ app/(dashboard)/dashboard/billing/_components/Billing.tsx
@@ -1,0 +1,69 @@
+import { redirect } from 'next/navigation';
+
+import { BillingForm } from '@/components/billing-form';
+import { DashboardHeader } from '@/components/header';
+import { Icons } from '@/components/icons';
+import { DashboardShell } from '@/components/shell';
+import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
+import { stripe } from '@/lib/stripe';
+import { getUserSubscriptionPlan } from '@/lib/subscription';
+
+const Billing = async () => {
+  const user = {
+    id: '123456',
+    name: 'manikanta',
+    image: './vercel.svg',
+    email: 'manikanta.potnuru@resonateaes.com',
+  };
+
+  if (!user) {
+    redirect('/login');
+  }
+
+  const subscriptionPlan = await getUserSubscriptionPlan(user.id);
+
+  // If user has a pro plan, check cancel status on Stripe.
+  let isCanceled = false;
+  if (subscriptionPlan.isPro && subscriptionPlan.stripeSubscriptionId) {
+    const stripePlan = await stripe.subscriptions.retrieve(
+      subscriptionPlan.stripeSubscriptionId
+    );
+    isCanceled = stripePlan.cancel_at_period_end;
+  }
+
+  return (
+    <DashboardShell>
+      <DashboardHeader
+        heading='Billing'
+        text='Manage billing and your subscription plan.'
+      />
+      <div className='grid gap-8'>
+        <Alert className='!pl-14'>
+          <Icons.warning />
+          <AlertTitle>This is a demo app.</AlertTitle>
+          <AlertDescription>
+            Taxonomy app is a demo app using a Stripe test environment. You can
+            find a list of test card numbers on the{' '}
+            <a
+              href='https://stripe.com/docs/testing#cards'
+              target='_blank'
+              rel='noreferrer'
+              className='font-medium underline underline-offset-8'
+            >
+              Stripe docs
+            </a>
+            .
+          </AlertDescription>
+        </Alert>
+        <BillingForm
+          subscriptionPlan={{
+            ...subscriptionPlan,
+            isCanceled,
+          }}
+        />
+      </div>
+    </DashboardShell>
+  );
+};
+
+export default Billing;
