+++ app/(dashboard)/dashboard/page.tsx
@@ -1,9 +1,9 @@
-import TodoView from '@/app/_views/TodoView';
+import Dashboard from './_components/Dashboard';
 
 export const metadata = {
   title: 'Dashboard',
 };
 
 export default function DashboardPage() {
-  return <TodoView />;
+  return <Dashboard />;
 }
