+++ app/(dashboard)/dashboard/settings/_components/Settings.tsx
@@ -1,0 +1,10 @@
+'use client';
+
+import SettingsView from '@/app/_views/SettingsView';
+import isAuth from '@/components/auth/is-auth';
+
+const Settings = () => {
+  return <SettingsView />;
+};
+
+export default isAuth(Settings);
