+++ app/_views/RegisterView.tsx
@@ -1,5 +1,3 @@
-'use client';
-
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import Link from 'next/link';
 import { useLayoutEffect, useState } from 'react';
